+++ flinkboot-core/src/test/java/io/github/sekelenao/flinkboot/core/internal/startup/CommandLineTest.java
@@ -1,10 +1,9 @@
 package io.github.sekelenao.flinkboot.core.internal.startup;
 
+import io.github.sekelenao.flinkboot.core.api.exception.parsing.CommandLineParsingException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-
-import java.util.NoSuchElementException;
 
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -58,10 +57,11 @@ class CommandLineTest {
         }
 
         @Test
-        @DisplayName("Should throw NoSuchElementException when option is missing its value")
+        @DisplayName("Should throw CommandLineParsingException when option is missing its value")
         void shouldThrowExceptionWhenOptionMissingValue() {
             String[] args = {"-key"};
-            assertThrows(NoSuchElementException.class, () -> CommandLine.parse(args));
+            var exception = assertThrows(CommandLineParsingException.class, () -> CommandLine.parse(args));
+            assertEquals("Option '-key' requires a value.", exception.getMessage());
         }
 
         @Test